requirements.txt
numpy==2.0.1
pandas==2.2.2
plotly==5.23.0
PyYAML==6.0.1
PyYAML==6.0.1
scipy==1.14.0
streamlit==1.36.0
streamlit_authenticator==0.2.2
streamlit_extras==0.4.3
